Use exact servertime on client side
As it is all singleplayer, there doesnt seem to be much reason for
handling the difference between client and server time as usual for
online play. Instead we just use the exact server time. That fixes
several stuttering issues that can happen when using slowmo timescales
(e.g. during force speed), pausing the game (e.g. opening inventory menu
and so on. Especially useful for segmented to avoid stuttering after
loading a save for the next segment.