Fix begin mission button in weapon select
Function UI_WeaponAllocBeginButton assumes the beginmission buttons to
be in increasing order with the weapon_menu cvar that they are shown
for. Unfortunately when adding the extra ones for custom mission
selection of later parts of story missions, this was overlooked. That
caused the wrong button to be activated/deactivated. This fixes that by
reordering the buttons properly according to the weapon_menu cvar they
are shown for.