Fix save/load abuse of speedrun timer
Saving and immediately loading that save can cause the player to be put
back into the server at a later servertime than when the load was
issued. As the timer is paused during this, that basically means that
you would teleport into the future without this time being accounted for
in the speedrun timer. When waiting for any timed events, this could be
abused to "save time" on the speedrun timer.

This is clearly not desirable, so now we add exactly this time delta of
jumps into the future to the speedrun timer.